+++ lecture-planner-graphql/src/main/java/com/github/mitschi/resolvers/Query.java
@@ -6,11 +6,10 @@ import com.github.mitschi.models.Lecture;
 import com.github.mitschi.services.EmployeeService;
 import com.github.mitschi.services.LectureService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
 import org.springframework.stereotype.Component;
 
+import java.util.Arrays;
 import java.util.List;
-import java.util.UUID;
 
 @Component
 public class Query implements GraphQLQueryResolver {
@@ -20,7 +19,7 @@ public class Query implements GraphQLQueryResolver {
     @Autowired
     private LectureService lectureService;
 
-    public Lecture lectureById(Integer id){
+    public Lecture lectureById(long id){
         return lectureService.getLectureById(id);
     }
